requirements.txt
streamlit
Flask>=2.3.0
crewai>=0.40.0
crewai-tools>=0.5.0
openai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
streamlit-cookies-manager>=0.2.0
pandas>=2.2.2
